support replication entry glob filters
